Add operations for inserting weight and length measurements.
Fixes #1.

+++ src/main/kotlin/org/vonvikken/growthbot/db/Measure.kt
@@ -8,7 +8,7 @@ import java.time.LocalDate
 
 internal sealed class Measure(name: String) : IntIdTable(name = name) {
     val babyID = reference(name = "baby_id", refColumn = id)
-    val measure: Column<Double> = double("measure")
+    val measure: Column<Int> = integer("measure")
     val date: Column<LocalDate> = date("date")
     val percentile: Column<MeasureRange.Percentile> = enumeration("percentile", MeasureRange.Percentile::class)
 }
